requirements-dev.txt
# Development-specific requirements

# import production requirements
-r requirements.txt

# development requirements
pycodestyle==2.10.0
coverage==7.0.5
Pillow==9.4.0
pytest-django==4.5.2
mixer==7.2.2